Full working multiprocessing pipeline

See merge request adaalman/AdaptiveHAC!1

+++ AdaptiveHAC/pointTransformer/train_cls.py
@@ -5,29 +5,36 @@ modified by zhongyuan
 from AdaptiveHAC.pointTransformer.dataset import ModelNetDataLoader, PCModelNetDataLoader, PCFileModelNetDataLoader
 import numpy as np
 import os
-os.environ['CUDA_LAUNCH_BLOCKING'] = "1"
 import torch
 import logging
 from tqdm import tqdm
 from AdaptiveHAC.pointTransformer import provider
-from AdaptiveHAC.processing.PointCloud import PointCloud
 import importlib
 import shutil
 import hydra
 import omegaconf
 import matplotlib.pyplot as plt
-import argparse
+import pickle
 from sklearn.model_selection import train_test_split
+from sklearn import metrics
+from AdaptiveHAC.pointTransformer import point_transformer
 
 device = torch.device("cuda:0" if torch.cuda.is_available() else "cpu")
 path=os.getcwd()
 # dataset = 'MMA_xyzI'
-
+def sanitiser(dataset):
+    for PC in dataset:
+            PC.data[np.isinf(PC.data)] = 1
+            PC.data[np.isneginf(PC.data)] = 0
+            PC.data[np.isnan(PC.data)] = 0
+    
+    return dataset
+            
 def validate(model, loader, num_class): #num_class should change !!!
     mean_correct = []
     class_acc = np.zeros((num_class,3))
     for j, data in tqdm(enumerate(loader), total=len(loader)):
-        points, target = data
+        points, target, _, _, _, _, _, _ = data
         target = target[:, 0]
         points, target = points.to(device), target.to(device)
         classifier = model.eval()
@@ -44,8 +51,7 @@ def validate(model, loader, num_class): #num_class should change !!!
     instance_acc = np.mean(mean_correct)
     return instance_acc, class_acc
 
-#@hydra.main(config_path='config', config_name='cls', version_base='1.1')
-def main_PC(args):
+def main(args):
     if isinstance(args, list):
         PC = args[1]
         args = args[0]
@@ -65,15 +71,24 @@ def main_PC(args):
     '''HYPER PARAMETER'''
     os.environ["CUDA_VISIBLE_DEVICES"] = str(args.gpu)
     logger = logging.getLogger(__name__)
+    logging.basicConfig()
 
     '''DATA LOADING'''
     logger.info('Load dataset ...')
             
     if PC != None:
         if PC_type == "single_node":
-            TRAIN_PC, TEST_PC = train_test_split(PC, train_size=0.8, shuffle=True)
+            TRAIN_PC, TEST_PC = train_test_split(PC, train_size=0.8, shuffle=True, random_state=1)
             TRAIN_DATASET = PCModelNetDataLoader(PC=TRAIN_PC, npoint=args.num_point)
             TEST_DATASET = PCModelNetDataLoader(PC=TEST_PC, npoint=args.num_point)
+            mean_label_class  = []
+            mean_labels = []
+            for i in PC:
+                mean_labels.append(i.mean_label) 
+            for j in range(1, 9, 1): #loop through classes 
+                mean_label_class.append(mean_labels.count(j))
+                
+            logger.info(f'Train_PC_len: {len(TRAIN_PC)} Test_PC_len: {len(TEST_PC)}')
         elif PC_type == "multiple_nodes":
             TRAIN_PC, TEST_PC = train_test_split(PC, train_size=0.8, shuffle=True)
             PC_TRAIN_all = []
@@ -84,25 +99,77 @@ def main_PC(args):
                 PC_TEST_all.extend(j)               
             TRAIN_DATASET = PCModelNetDataLoader(PC=PC_TRAIN_all, npoint=args.num_point)
             TEST_DATASET = PCModelNetDataLoader(PC=PC_TEST_all, npoint=args.num_point)
+            mean_label_class  = []
+            mean_labels = []
+            for i in PC:
+                for node in i:
+                    mean_labels.append(node.mean_label) 
+            for j in range(1, 10, 1): #loop through classes 
+                mean_label_class.append(mean_labels.count(j) if mean_labels.count(j) != 0 else 1)
+       	    logger.info(f'Train_PC_len: {len(PC_TRAIN_all)} Test_PC_len: {len(PC_TEST_all)}')
     else:
         dataset = args.dataset
         DATA_PATH = hydra.utils.to_absolute_path(dataset)
         TRAIN_DATASET = ModelNetDataLoader(root=DATA_PATH, npoint=args.num_point, split='train')
         TEST_DATASET = ModelNetDataLoader(root=DATA_PATH, npoint=args.num_point, split='test')
+    
+    PC_TRAIN_all = sanitiser(PC_TRAIN_all)
+    PC_TEST_all = sanitiser(PC_TEST_all)
         
+    with open('TRAIN_PC.pkl', 'wb') as file:
+            pickle.dump(TRAIN_PC, file)
+    with open('TEST_PC.pkl', 'wb') as file:
+            pickle.dump(TEST_PC, file)    
+            
     trainDataLoader = torch.utils.data.DataLoader(TRAIN_DATASET, batch_size=args.batch_size, shuffle=True, num_workers=4)
     testDataLoader = torch.utils.data.DataLoader(TEST_DATASET, batch_size=args.batch_size, shuffle=False, num_workers=4)
 
     '''MODEL LOADING'''
-    shutil.copy(hydra.utils.to_absolute_path('pointTransformer/models/{}/model.py'.format(args.model.name)), '.')
-
-    classifier = getattr(importlib.import_module('pointTransformer.models.{}.model'.format(args.model.name)), 'PointTransformerCls')(args)
+    if os.path.exists(hydra.utils.to_absolute_path('pointTransformer/models/{}/model.py'.format(args.model.name))): # for running with processing
+        shutil.copy(hydra.utils.to_absolute_path('pointTransformer/models/{}/model.py'.format(args.model.name)), '.')
+        classifier = getattr(importlib.import_module('pointTransformer.models.{}.model'.format(args.model.name)), 'PointTransformerCls')(args)
+    elif os.path.exists(os.path.abspath('../../../../pointTransformer/models/{}/model.py'.format(args.model.name))): # for only training
+        shutil.copy(os.path.abspath('../../../../pointTransformer/models/{}/model.py'.format(args.model.name)), '.')
+        classifier = getattr(importlib.import_module('models.{}.model'.format(args.model.name)), 'PointTransformerCls')(args)
+    
     if torch.cuda.device_count() > 1:
         classifier = torch.nn.DataParallel(classifier)
     classifier = classifier.to(device)
-    # print(classifier)
-    criterion = torch.nn.CrossEntropyLoss(torch.FloatTensor([0,1,1,1,1,1,1,1,1,1]).to(device))
+    
+    match args.loss_function:
+        case "Custom": 
+            weights = []
+            # weights options
+            weight_option = 2
+            match weight_option:
+                case 1: # good result
+                    for j in range(0, 9, 1):
+                        weights.append(1/(mean_label_class[j]))
+                    weights = np.asarray(weights)/sum(weights)
+                case 2: # good result
+                    for j in range(0, 9, 1):
+                        weights.append(1/(mean_label_class[j]))
+                case 3: # good result
+                    for j in range(0, 9, 1):
+                        weights.append(1/(mean_label_class[j]/sum(mean_label_class)))
+                case 4: 
+                    for j in range(0, 9, 1):
+                        weights.append((1/(mean_label_class[j]))**2)
+                    weights = np.array(weights)/sum(weights)
 
+            criterion = torch.nn.CrossEntropyLoss(torch.FloatTensor([0,
+                                                                    weights[0],
+                                                                    weights[1],
+                                                                    weights[2],
+                                                                    weights[3],
+                                                                    weights[4],
+                                                                    weights[5],
+                                                                    weights[6],
+                                                                    weights[7],
+                                                                    weights[8]]).to(device))
+        case "Default":
+            criterion = torch.nn.CrossEntropyLoss(torch.FloatTensor([0,1,1,1,1,1,1,1,1,1]).to(device))
+    
     try:
         checkpoint = torch.load('best_model.pth')
         start_epoch = checkpoint['epoch']
@@ -143,7 +210,7 @@ def main_PC(args):
         logger.info('Epoch %d (%d/%s):' % (global_epoch + 1, epoch + 1, args.epoch))
         classifier.train()
         for batch_id, data in tqdm(enumerate(trainDataLoader, 0), total=len(trainDataLoader), smoothing=0.9):
-            points, target = data
+            points, target, _, _, _, _, _, _ = data
             points = points.data.numpy()
             points = provider.random_point_dropout(points)
             points[:,:, 0:3] = provider.random_scale_point_cloud(points[:,:, 0:3])
@@ -226,11 +293,11 @@ def main_PC(args):
             f.write(' ')
         
         f.write('\n')
-        for acc in test_accuracy :
+        for acc in test_accuracy:
             f.write(str(acc))
             f.write(' ')
     logger.info('End of savetxt...')
-    
+    torch.cuda.empty_cache()
     return TEST_PC, classifier.eval()
 
 def main(args):
@@ -423,4 +490,26 @@ def main(args):
     return TEST_PC, classifier.eval()
     
 if __name__ == '__main__':
-    main()
+    hydra.initialize(config_path="config", version_base='1.3')
+    args = hydra.compose(config_name='cls', return_hydra_config=True)
+    omegaconf.OmegaConf.set_struct(args, False)
+    args.model = args._group_
+    args.pop('_group_')
+    
+    logger = logging.getLogger(__name__)
+    logging.basicConfig()
+    
+    os.makedirs(os.path.abspath(args.hydra.run.dir))
+    os.chdir(os.path.abspath(args.hydra.run.dir))
+    #path = "../test"
+    path = "C:/Users/adaal/OneDrive - Delft University of Technology/Internship HAC/Results/ExtraRuns/lagsearchTH03CustomLoss_2"
+    with open(f'{path}/Processed_data.pkl', 'rb') as file:
+        PC_dataset = pickle.load(file)
+    TEST_PC, model = main([args, PC_dataset])
+    logger.info("Testing on dataset...")
+    F1_scores, acc, balanced_acc = point_transformer.test(args, model, 'softmax', TEST_PC)
+        
+    if 'softmax' != "none":
+        logger.info(f"Fused: F1 score: {F1_scores}, accuracy: {acc}, balanced accuracy: {balanced_acc}")
+    else: 
+        logger.info("\n".join([f"Node {i}: F1 score: {F1_scores[i]}, accuracy: {acc[i]}, balanced accuracy: {balanced_acc[i]}" for i in range(len(F1_scores))]))
